feat(user): add user feature
- Add getUser API function
- Remove legacy user modules
