--- src/user/UserProvider.tsx
@@ -1,8 +1,0 @@
-import * as React from 'react';
-import { useSyncUser } from './useSyncUser';
-
-export const UserProvider = React.memo(() => {
-  useSyncUser();
-
-  return null;
-});
